% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgatlw.R
\docType{data}
\name{kgatlw_pal}
\alias{kgatlw_pal}
\title{King Gizzard and the Lizard Wizard palettes}
\format{An object of class \code{list} of length 18.}
\usage{
kgatlw_pal
}
\description{
King Gizzard and the Lizard Wizard palettes
}
\keyword{datasets}
