% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kgatlw.R
\name{kgatlw_palette}
\alias{kgatlw_palette}
\title{King Gizzard and the Lizard Wizard palette generator}
\usage{
kgatlw_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of palette. List of available palettes:
\code{twelvebar}, \code{elts}, \code{floatalong}, \code{oddments},
\code{mindfuzz}, \code{quarters1}, \code{papermache}, \code{nonagon},
\code{microbanana}, \code{motu}, \code{sketches}, \code{polygondwana},
\code{gumboot}, \code{gizzfest2017}, \code{gizzfest2018}, \code{mex2018},
\code{fff}, \code{itrn}}

\item{n}{Number of colors. Number of colors available depends on palette, ranging from three colors
(microbanana) to eight (quarters1, gizzfest2017, fff, and itrn).}

\item{type}{Choose between "discrete" or "continuous"}
}
\value{
A vector of colors.
}
\description{
Allows you to use palettes inspired by King Gizzard's artwork by Jason Galea
}
\examples{
kgatlw_palette("microbanana")
kgatlw_palette("oddments", 4)
kgatlw_palette("polygondwana", n = 20, type = "continuous")

}
